backend/python/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
psycopg2-binary
pymongo